feat/tests: add tests and implement delete account route

+++ src/service/DeleteUser.ts
@@ -1,16 +1,25 @@
 import type { UserRepository } from '../repositories/User';
 import { BadRequestError } from '../config/BaseError';
-import { NON_EXISTENT_USER } from '../utils/messages';
+import { NON_EXISTENT_USER, PASSWORD_INCORRECT } from '../utils/messages';
 import { UUID } from '../@types';
+import bcrypt from 'bcrypt';
 
 export class DeleteUserService {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async execute(userId: UUID) {
-    const numberOfRowDeleted = await this.userRepository.deleteUser(userId);
+  async execute(userId: UUID, password: string) {
+    const userRegister = await this.userRepository.getUserById(userId);
 
-    if (!numberOfRowDeleted) {
+    if (!userRegister) {
       throw new BadRequestError(NON_EXISTENT_USER);
+    }
+
+    const checkPassword = await bcrypt.compare(password, userRegister.password_hash);
+
+    if (checkPassword) {
+      await this.userRepository.deleteUser(userId);
+    } else {
+      throw new BadRequestError(PASSWORD_INCORRECT);
     }
   }
 }
